test_missingNumber.py
class Solution(object):
    def missingNumber(self, nums):
        nums.sort()
        for key, value in enumerate(nums):
            if key != value:
                return key


# Test cases
# Test case 1
nums = [3, 0, 1]
expected_output = 2
assert Solution().missingNumber(nums) == expected_output, "Test case 1 failed"

# Test case 2
nums = [0, 1]
expected_output = 2
assert Solution().missingNumber(nums) == expected_output, "Test case 2 failed"

# Test case 3
nums = [9, 6, 4, 2, 3, 5, 7, 0, 1]
expected_output = 8
assert Solution().missingNumber(nums) == expected_output, "Test case 3 failed"

# Remove the following lines as they are not necessary for the code execution
# Example 1:
# Input: nums = [3,0,1]
# Output: 2
# Explanation: n = 3 since there are 3 numbers, so all numbers are in the range [0,3]. 2 is the missing number in the range since it does not appear in nums.

# Example 2:
# Input: nums = [0,1]
# Output: 2
# Explanation: n = 2 since there are 2 numbers, so all numbers are in the range [0,2]. 2 is the missing number in the range since it does not appear in nums.

# Example 3:
# Input: nums = [9,6,4,2,3,5,7,0,1]
# Output: 8
# Explanation: n = 9 since there are 9 numbers, so all numbers are in the range [0,9]. 8 is the missing number in the range since it does not appear in nums.
